% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aether_wgcna_plots.R
\name{AETHER_plot_sample_dendrogram}
\alias{AETHER_plot_sample_dendrogram}
\title{Plot sample dendrogram with trait colors}
\usage{
AETHER_plot_sample_dendrogram(
  cluster_result,
  traits = NULL,
  main = "Sample dendrogram and trait heatmap",
  ...
)
}
\arguments{
\item{cluster_result}{A wgcna_cluster object from ARTEMIS_wgcna_cluster_samples(),
or an hclust object.}

\item{traits}{Data.frame of traits (samples as rows) or wgcna_data object.}

\item{main}{Plot title.}

\item{...}{Additional arguments passed to plotDendroAndColors().}
}
\value{
Invisible NULL. Plots to current device.
}
\description{
Creates a sample clustering dendrogram with trait values shown as colored
bars underneath.
}
\examples{
\dontrun{
clust <- ARTEMIS_wgcna_cluster_samples(wgcna_data)
AETHER_plot_sample_dendrogram(clust, wgcna_data)

}
}
