% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agprodchange_ni_gcam6}
\alias{agprodchange_ni_gcam6}
\title{agprodchange_ni_gcam6}
\format{
R data frame
}
\source{
GCAM6 ag productivity change reference data https://jgcri.github.io/gcam-doc/inputs_supply.html
}
\usage{
agprodchange_ni_gcam6
}
\description{
agprodchange_ni_gcam6
}
\examples{
\dontrun{
 library(gaia);
 agprodchange_ni_gcam6 <- gaia::agprodchange_ni_gcam6
}
}
\keyword{datasets}
