% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agprodchange_ni_gcam7}
\alias{agprodchange_ni_gcam7}
\title{agprodchange_ni_gcam7}
\format{
R data frame
}
\source{
GCAM7 ag productivity change reference data
}
\usage{
agprodchange_ni_gcam7
}
\description{
agprodchange_ni_gcam7
}
\examples{
\dontrun{
 library(gaea);
 agprodchange_ni_gcam7 <- gaea::agprodchange_ni_gcam7
}
}
\keyword{datasets}
