% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_utils.R
\name{agprodchange_ref}
\alias{agprodchange_ref}
\title{agprodchange_ref}
\usage{
agprodchange_ref(
  gcam_version = "gcam7",
  gcam_timestep = 5,
  base_year = 2015,
  climate_scenario = NULL,
  gcamdata_dir = NULL
)
}
\arguments{
\item{gcam_version}{Default = 'gcam7'. String for the GCAM version. Only support gcam6 and gcam7}

\item{gcam_timestep}{Default = 5. Integer for the time step of GCAM (Select either 1 or 5 years for GCAM use)}

\item{base_year}{Default = 2015. Integer for the base year (for GCAM)}

\item{climate_scenario}{Default = NULL. String for climate scenario (e.g., 'ssp245')}

\item{gcamdata_dir}{Default = NULL. String for directory to the gcamdata folder within the specific GCAM version. The gcamdata need to be run with drake to have the CSV outputs beforehand.}
}
\value{
A data frame of the baseline agricultural productivity change from GCAM
}
\description{
Get the reference agricultural productivity change based GCAM version
}
\keyword{internal}
