% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_peak_annotation.R
\name{APOLLO_enrich_kegg}
\alias{APOLLO_enrich_kegg}
\title{KEGG Pathway Enrichment Analysis}
\usage{
APOLLO_enrich_kegg(
  annotated_peaks,
  org_db = "org.Hs.eg.db",
  organism = "hsa",
  gene_id_type = "ENTREZID",
  pval_cutoff = 0.05,
  qval_cutoff = 0.1,
  verbose = TRUE
)
}
\arguments{
\item{annotated_peaks}{Data.frame from APOLLO_annotate_peaks() with gene_id column.
Alternatively, a character vector of gene IDs.}

\item{org_db}{Character string or OrgDb object. The organism annotation database
for ID conversion. Default is "org.Hs.eg.db" for human.}

\item{organism}{Character. KEGG organism code (default = "hsa" for human).}

\item{gene_id_type}{Character. Type of gene IDs provided. One of "ENTREZID",
"ENSEMBL", "SYMBOL", "REFSEQ" (default = "ENTREZID").}

\item{pval_cutoff}{Numeric. P-value cutoff for enrichment (default = 0.05).}

\item{qval_cutoff}{Numeric. Adjusted p-value cutoff (default = 0.1).}

\item{verbose}{Logical. Print summary (default = TRUE).}
}
\value{
An enrichResult object from clusterProfiler.
}
\description{
Performs KEGG pathway enrichment analysis on genes associated
with annotated peaks using clusterProfiler.
}
\examples{
\dontrun{
annotated <- APOLLO_annotate_peaks(sig_peaks, txdb)
kegg_results <- APOLLO_enrich_kegg(annotated)
dotplot(kegg_results)

}
}
