% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_activity_networks.R
\name{APOLLO_list_activity_databases}
\alias{APOLLO_list_activity_databases}
\title{List available decoupleR databases}
\usage{
APOLLO_list_activity_databases()
}
\value{
Invisibly returns a data.frame of database information.
}
\description{
Prints information about available prior knowledge databases
for activity inference.
}
\examples{
\dontrun{
APOLLO_list_activity_databases()

}
}
