% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artemis_decoupler.R
\name{ARTEMIS_list_methods}
\alias{ARTEMIS_list_methods}
\title{List available decoupleR methods}
\usage{
ARTEMIS_list_methods(recommended_only = FALSE)
}
\arguments{
\item{recommended_only}{Logical. If TRUE, only show recommended methods.
Default: FALSE.}
}
\value{
Character vector of method names (invisibly).
}
\description{
Prints information about available statistical methods for activity inference.
}
\examples{
\dontrun{
ARTEMIS_list_methods()
ARTEMIS_list_methods(recommended_only = TRUE)

}
}
