% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artemis_part_clustering.R
\name{ARTEMIS_part_summary}
\alias{ARTEMIS_part_summary}
\title{Print PART result summary}
\usage{
ARTEMIS_part_summary(part_result, verbose = TRUE)
}
\arguments{
\item{part_result}{An \code{artemis_part} object.}

\item{verbose}{Logical. Print to console. Default: TRUE.}
}
\value{
Invisible NULL.
}
\description{
Print PART result summary
}
