% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artemis_wgcna.R
\name{ARTEMIS_wgcna_module_traits}
\alias{ARTEMIS_wgcna_module_traits}
\title{Correlate modules with traits}
\usage{
ARTEMIS_wgcna_module_traits(
  modules,
  traits,
  cor_method = "pearson",
  p_adjust = "BH",
  verbose = TRUE
)
}
\arguments{
\item{modules}{A wgcna_modules object from ARTEMIS_wgcna_detect_modules().}

\item{traits}{Data.frame of traits, or wgcna_data object (uses datTraits).
Must have samples as rows matching the module data.}

\item{cor_method}{Correlation method: "pearson" or "spearman". Default: "pearson".}

\item{p_adjust}{Method for p-value adjustment: "BH", "bonferroni", "none", etc.
Default: "BH" (Benjamini-Hochberg).}

\item{verbose}{Logical. Print progress messages. Default: TRUE.}
}
\value{
A list with class "wgcna_trait_cor" containing:
\item{cor_matrix}{Matrix of correlations (modules x traits)}
\item{pvalue_matrix}{Matrix of p-values}
\item{padj_matrix}{Matrix of adjusted p-values}
\item{significant_associations}{Data.frame of significant module-trait pairs}
\item{n_significant}{Number of significant associations}
}
\description{
Calculates correlations between module eigengenes and sample traits.
Identifies which modules are associated with which clinical/phenotypic variables.
}
\examples{
\dontrun{
modules <- ARTEMIS_wgcna_detect_modules(wgcna_data, power = 6)
trait_cor <- ARTEMIS_wgcna_module_traits(modules, wgcna_data)

}
}
