% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_legal}
\alias{check_legal}
\title{Check table format and table name.}
\usage{
check_legal(format, table_name)
}
\arguments{
\item{format}{output table format}

\item{table_name}{table name}
}
\description{
Check table format and table name.
}
