% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_calendars.R
\name{clean_sage}
\alias{clean_sage}
\title{clean_sage}
\usage{
clean_sage(d = NULL, crop_select = NULL)
}
\arguments{
\item{d}{Default = NULL. Data frame for crop calendar}

\item{crop_select}{Default = NULL. Vector of strings for selected crops}
}
\value{
A data frame of clean crop calendar
}
\description{
Clean up the sage crop calendar by reassigning the iso name and
make sure to keep one set of the plant and harvest for each country and crop
}
\keyword{internal}
