% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_impacts_functions.R
\name{clean_yield}
\alias{clean_yield}
\title{clean_yield}
\usage{
clean_yield(fao_yield = NULL)
}
\arguments{
\item{fao_yieldDefault}{= NULL.  data frame for the fao yield table}
}
\description{
Function to clean FAO yield data
}
\keyword{internal}
