% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_impacts_functions.R
\name{climate_impact}
\alias{climate_impact}
\title{climate_impact}
\usage{
climate_impact(
  climate_model = NULL,
  climate_scenario = NULL,
  crop_name = NULL,
  base_year = NULL,
  start_year = NULL,
  end_year = NULL,
  output_dir = NULL
)
}
\arguments{
\item{climate_model}{Default = NULL. string for climate model name}

\item{climate_scenario}{Default = NULL. string for RCP}

\item{crop_name}{Default = NULL. string for crop name}

\item{base_year}{Default = NULL. integer for the base year (for GCAM)}

\item{start_year}{Default = NULL. integer for the  start year of the data}

\item{end_year}{Default = NULL. integer for the end year of the data}

\item{output_dir}{Default = NULL. string for path to the output folder}
}
\description{
Estimate climate impacts (ratio future to each base year, then average)
}
\keyword{internal}
