% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{co2_historical}
\alias{co2_historical}
\title{co2_historical}
\format{
R data frame
}
\source{
Historical CO2 concentration hist_co2_conc.csv
}
\usage{
co2_historical
}
\description{
co2_historical
}
\examples{
\dontrun{
 library(gaia);
 co2_historical <- gaia::co2_historical
}
}
\keyword{datasets}
