% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coef_default}
\alias{coef_default}
\title{coef_default}
\format{
R data frame
}
\source{
Regression coefficients derived by using the default historical WATCH climate data
}
\usage{
coef_default
}
\description{
Regression coefficients generated by \href{https://broom.tidymodels.org/}{`broom::tidy()`} for the regression model using gaia's example historical climate data, including estimate, standard error, statistic, p-value, for the 17 default crops supported by gaia.
}
\examples{
\dontrun{
 library(gaia);
 coef_default <- gaia::coef_default
}
}
\keyword{datasets}
