% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_id}
\alias{country_id}
\title{country_id}
\format{
R data frame
}
\source{
country ID and name mapping
}
\usage{
country_id
}
\description{
country_id
}
\examples{
\dontrun{
 library(gaia);
 country_id <- gaia::country_id
}
}
\keyword{datasets}
