\name{crc_markers}
\alias{crc_markers}
\docType{data}
\title{Markers Descriptors for Real aCGH Dataset of Colorectal Cancer (CRC)}
\description{
Markers Descriptors of CRC dataset for SNP 250k Affymetrix GeneChip arrays.
}
\author{Sandro Morganella et al.

Maintainer: S. Morganella <morganellaalx@gmail.com>
}
\usage{data(crc_markers)}

\format{
	The marker descriptor matrix is organized as a matrix having a row for each measured probe. Each probe is described by the following columns:\cr
	Probe Name - Chromosome - Start\cr
}
\examples{
data(crc_markers)

}
\keyword{datasets}
