% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_calendars.R
\name{crop_calendars}
\alias{crop_calendars}
\title{crop_calendars}
\usage{
crop_calendars(output_dir = file.path(getwd(), "output"))
}
\arguments{
\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
}
\description{
Generate planting months for each country
Data from SAGE
}
