% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aggregation.R
\name{data_aggregation}
\alias{data_aggregation}
\title{data_aggregation}
\usage{
data_aggregation(
  climate_hist_dir = NULL,
  climate_impact_dir = NULL,
  climate_model = "gcm",
  climate_scenario = "rcp",
  output_dir = file.path(getwd(), "output")
)
}
\arguments{
\item{climate_hist_dir}{Default = NULL. string for path to the processed historical climate data folder}

\item{climate_impact_dir}{Default = NULL. string for path to the processed future climate data folder using weighted_climate function}

\item{climate_model}{Default = NULL. string for climate model (e.g., 'CanESM5')}

\item{climate_scenario}{Default = NULL. string for climate scenario (e.g., 'ssp245')}

\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
}
\description{
Process multiple models to analyze historical weather variables and crop yields
}
