% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_impacts_functions.R
\name{data_merge}
\alias{data_merge}
\title{data_merge}
\usage{
data_merge(
  data = NULL,
  crop_name = NULL,
  yield = NULL,
  output_dir = NULL,
  co2_hist = NULL,
  gdp_hist = gdp
)
}
\arguments{
\item{data}{Default = NULL. data table for crop data created by crop_month}

\item{crop_name}{Default = NULL. string for crop name}

\item{yield}{Default = NULL. data table for yield data}

\item{output_dir}{Default = NULL. string for path to output folder}

\item{co2_hist}{Default = NULL data table for historical CO2 concentration [year, co2_conc]}

\item{gdp_hist}{Default = gdp. data table for historical GDP by ISO [iso, year, gdp_pcap_ppp]}
}
\description{
Function to merge the iso codes, FAO yield (historic), crop calendar, and weather variables
and select variables for regression analysis
}
\keyword{internal}
