% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demeter_result_loaders.R
\name{DEMETER_filter_circos_clusters}
\alias{DEMETER_filter_circos_clusters}
\title{Filter circos data to specific PART clusters}
\usage{
DEMETER_filter_circos_clusters(circos_data, selected_clusters)
}
\arguments{
\item{circos_data}{A \code{part_wgcna_circos} object.}

\item{selected_clusters}{Character vector of cluster names to keep.}
}
\value{
Updated \code{part_wgcna_circos} object.
}
\description{
Filter circos data to specific PART clusters
}
