% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demeter_matrix_utils.R
\name{DEMETER_matrix_power}
\alias{DEMETER_matrix_power}
\title{Matrix Power Function (Function Form)}
\usage{
DEMETER_matrix_power(x, n)
}
\arguments{
\item{x}{A square matrix}

\item{n}{Power to raise the matrix to (can be negative for inverse powers)}
}
\value{
Matrix x raised to the power n
}
\description{
Raises a matrix to a given power using eigenvalue decomposition.
Function form of the \%^\% operator for explicit module referencing.
}
\examples{
\dontrun{
# Compute inverse square root of a covariance matrix
Sigma <- cov(matrix(rnorm(100), 10, 10))
Sigma_inv_sqrt <- DEMETER_matrix_power(Sigma, -0.5)

}
}
