% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_isoform_annotation.R
\name{.apollo_cpat_run}
\alias{.apollo_cpat_run}
\title{Execute CPAT inside the basilisk environment}
\usage{
.apollo_cpat_run(cli_args, verbose)
}
\description{
Execute CPAT inside the basilisk environment
}
\keyword{internal}
