% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_isoform_annotation.R
\name{.apollo_deeploc2_run}
\alias{.apollo_deeploc2_run}
\title{Execute DeepLoc 2.1 inside the basilisk environment}
\usage{
.apollo_deeploc2_run(fasta, output_dir, model, verbose)
}
\description{
Execute DeepLoc 2.1 inside the basilisk environment
}
\keyword{internal}
