% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_isoform_annotation.R
\name{.apollo_deeptmhmm_run}
\alias{.apollo_deeptmhmm_run}
\title{Execute DeepTMHMM's predict.py inside the basilisk environment}
\usage{
.apollo_deeptmhmm_run(deeptmhmm_dir, fasta, output_dir, verbose)
}
\description{
Execute DeepTMHMM's predict.py inside the basilisk environment
}
\keyword{internal}
