% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_isoform_annotation.R
\name{.apollo_iupred2a_run}
\alias{.apollo_iupred2a_run}
\title{Execute IUPred2A once per isoform inside the basilisk environment}
\usage{
.apollo_iupred2a_run(script_path, sequences, split_dir, verbose)
}
\description{
Execute IUPred2A once per isoform inside the basilisk environment
}
\keyword{internal}
