% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_isoform_annotation.R
\name{.apollo_pfam_run}
\alias{.apollo_pfam_run}
\title{Execute pfam_scan.pl inside the basilisk environment}
\usage{
.apollo_pfam_run(cli_args, verbose)
}
\description{
Execute pfam_scan.pl inside the basilisk environment
}
\keyword{internal}
