% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_isoform_annotation.R
\name{.apollo_reformat_cpat}
\alias{.apollo_reformat_cpat}
\title{Reformat CPAT output to IsoformSwitchAnalyzeR's expected column layout}
\usage{
.apollo_reformat_cpat(raw_file, out_file)
}
\description{
Reformat CPAT output to IsoformSwitchAnalyzeR's expected column layout
}
\keyword{internal}
