% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_isoform_annotation.R
\name{.apollo_reformat_deeploc2}
\alias{.apollo_reformat_deeploc2}
\title{Reformat DeepLoc 2.1 output to IsoformSwitchAnalyzeR's expected column layout}
\usage{
.apollo_reformat_deeploc2(raw_file, out_file)
}
\description{
Reformat DeepLoc 2.1 output to IsoformSwitchAnalyzeR's expected column layout
}
\keyword{internal}
