% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_isoform_annotation.R
\name{.apollo_signalp_run}
\alias{.apollo_signalp_run}
\title{Execute SignalP 6 inside the basilisk environment}
\usage{
.apollo_signalp_run(fasta, output_dir, organism, mode, verbose)
}
\description{
Execute SignalP 6 inside the basilisk environment
}
\keyword{internal}
