% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aether_differential_plots.R
\name{.dea_extract_df}
\alias{.dea_extract_df}
\title{Extract data.frame from a DEA result object or data.frame}
\usage{
.dea_extract_df(dea_result, required)
}
\description{
Extract data.frame from a DEA result object or data.frame
}
\keyword{internal}
