% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aether_differential_plots.R
\name{.dea_resolve_label_col}
\alias{.dea_resolve_label_col}
\title{Resolve label column for DEA plots}
\usage{
.dea_resolve_label_col(df, label_col)
}
\description{
Resolve label column for DEA plots
}
\keyword{internal}
