% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hephaestus_MOFA.R
\name{.hephaestus_mofa_validate}
\alias{.hephaestus_mofa_validate}
\title{Validate list-of-matrices input for HEPHAESTUS_run_mofa()}
\usage{
.hephaestus_mofa_validate(X)
}
\description{
Validate list-of-matrices input for HEPHAESTUS_run_mofa()
}
\keyword{internal}
