% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_motif_enrichment.R
\name{.homer_build_batch}
\alias{.homer_build_batch}
\title{Build homer_motif_batch from a list of homer_motif results}
\usage{
.homer_build_batch(results, genome, bed_files)
}
\description{
Build homer_motif_batch from a list of homer_motif results
}
\keyword{internal}
