% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_motif_enrichment.R
\name{.homer_check_installation}
\alias{.homer_check_installation}
\title{Check HOMER installation}
\usage{
.homer_check_installation()
}
\value{
Path to findMotifsGenome.pl or stops with error
}
\description{
Check HOMER installation
}
\keyword{internal}
