% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_motif_enrichment.R
\name{.homer_parse_denovo}
\alias{.homer_parse_denovo}
\title{Parse HOMER de novo motif file}
\usage{
.homer_parse_denovo(motif_file)
}
\arguments{
\item{motif_file}{Path to homerMotifs.all.motifs}
}
\value{
data.frame with de novo motif summary
}
\description{
Parse HOMER de novo motif file
}
\keyword{internal}
