% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_motif_enrichment.R
\name{.homer_parse_known}
\alias{.homer_parse_known}
\title{Parse HOMER knownResults.txt}
\usage{
.homer_parse_known(known_file)
}
\arguments{
\item{known_file}{Path to knownResults.txt}
}
\value{
data.frame with cleaned columns
}
\description{
Parse HOMER knownResults.txt
}
\keyword{internal}
