% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_motif_enrichment.R
\name{.homer_rebuild_batch}
\alias{.homer_rebuild_batch}
\title{Rebuild batch combined/summary from individual results}
\usage{
.homer_rebuild_batch(batch)
}
\description{
Rebuild batch combined/summary from individual results
}
\keyword{internal}
