% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_motif_enrichment.R
\name{.homer_run}
\alias{.homer_run}
\title{Execute HOMER findMotifsGenome.pl inside the basilisk environment}
\usage{
.homer_run(cli_args, output_dir)
}
\value{
Exit code from findMotifsGenome.pl
}
\description{
Execute HOMER findMotifsGenome.pl inside the basilisk environment
}
\keyword{internal}
