% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aether_isoform_switch_plots.R
\name{.iso_extract_df}
\alias{.iso_extract_df}
\title{Extract per-isoform results data.frame from an artemis_isoform_switch object}
\usage{
.iso_extract_df(switch_result, required)
}
\description{
Extract per-isoform results data.frame from an artemis_isoform_switch object
}
\keyword{internal}
