% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aether_isoform_switch_plots.R
\name{.splicing_extract_df}
\alias{.splicing_extract_df}
\title{Extract a top-level data.frame field from an artemis_isoform_switch object}
\usage{
.splicing_extract_df(switch_result, field, from_fn)
}
\description{
Extract a top-level data.frame field from an artemis_isoform_switch object
}
\keyword{internal}
