% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artemis_clustering.R
\name{.test_quali_cluster}
\alias{.test_quali_cluster}
\title{Test qualitative variable against clusters}
\usage{
.test_quali_cluster(x, clusters)
}
\description{
Test qualitative variable against clusters
}
\keyword{internal}
