% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artemis_clustering.R
\name{.test_quanti_cluster}
\alias{.test_quanti_cluster}
\title{Test quantitative variable against clusters}
\usage{
.test_quanti_cluster(x, clusters)
}
\description{
Test quantitative variable against clusters
}
\keyword{internal}
