% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aether_network_plots.R
\name{.vis_continuous_colors}
\alias{.vis_continuous_colors}
\title{Map continuous values to a color gradient for visNetwork}
\usage{
.vis_continuous_colors(vals, colors = NULL)
}
\description{
Map continuous values to a color gradient for visNetwork
}
\keyword{internal}
