% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artemis_wgcna.R
\name{.wgcna_load_counts}
\alias{.wgcna_load_counts}
\title{Load counts from file or directory}
\usage{
.wgcna_load_counts(path, format, gene_col, verbose)
}
\description{
Load counts from file or directory
}
\keyword{internal}
