% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artemis_wgcna.R
\name{.wgcna_prepare_traits}
\alias{.wgcna_prepare_traits}
\title{Prepare traits data - convert to numeric where appropriate}
\usage{
.wgcna_prepare_traits(traits, verbose)
}
\description{
Prepare traits data - convert to numeric where appropriate
}
\keyword{internal}
