% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eleuthia_logging.R
\name{ELEUTHIA_get_log_file}
\alias{ELEUTHIA_get_log_file}
\title{Get Current Log File Path}
\usage{
ELEUTHIA_get_log_file()
}
\value{
Character string with log file path, or NULL if logging is inactive.
}
\description{
Returns the path to the currently active log file, or NULL
if no log is active.
}
\examples{
\dontrun{
ELEUTHIA_get_log_file()
# NULL

ELEUTHIA_start_log("my_analysis.log")
ELEUTHIA_get_log_file()
# "my_analysis.log"

ELEUTHIA_stop_log()

}
}
