% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eleuthia_logging.R
\name{ELEUTHIA_is_logging}
\alias{ELEUTHIA_is_logging}
\title{Check if Logging is Active}
\usage{
ELEUTHIA_is_logging()
}
\value{
Logical. TRUE if logging is active, FALSE otherwise.
}
\description{
Helper function to check if console logging is currently active.
}
\examples{
\dontrun{
ELEUTHIA_is_logging()
# FALSE

ELEUTHIA_start_log("test.log")
ELEUTHIA_is_logging()
# TRUE

ELEUTHIA_stop_log()
ELEUTHIA_is_logging()
# FALSE

}
}
