% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eleuthia_peak_loading.R
\name{ELEUTHIA_load_bed_from_sheet}
\alias{ELEUTHIA_load_bed_from_sheet}
\title{Load ChIP/CUT&TAG BED Files from Sample Sheet}
\usage{
ELEUTHIA_load_bed_from_sheet(sample_sheet, omics = "CHIPseq", verbose = TRUE)
}
\arguments{
\item{sample_sheet}{A validated sample sheet data.frame.}

\item{omics}{Character string. Omics type to load (default = "CHIPseq").}

\item{verbose}{Logical. Print progress messages (default = TRUE).}
}
\value{
A named list of BED data.frames, with names corresponding to sample_id.
}
\description{
Loads fragment BED files for ChIP-seq or CUT&TAG data
from a validated sample sheet.
}
