% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eleuthia_peak_loading.R
\name{ELEUTHIA_peaks_to_saf}
\alias{ELEUTHIA_peaks_to_saf}
\title{Convert Consensus Peaks to SAF Format}
\usage{
ELEUTHIA_peaks_to_saf(consensus_peaks)
}
\arguments{
\item{consensus_peaks}{A consensus peak data.frame from
ELEUTHIA_create_consensus_peaks().}
}
\value{
A data.frame in SAF format with columns: GeneID, Chr, Start, End, Strand.
}
\description{
Converts a consensus peak data.frame to SAF (Simplified
Annotation Format) for use with Rsubread::featureCounts().
}
\details{
SAF format is required by Rsubread::featureCounts() for counting reads
in custom genomic regions. The "GeneID" column contains the peak_id.
}
\examples{
\dontrun{
consensus <- ELEUTHIA_create_consensus_peaks(peak_list)
saf <- ELEUTHIA_peaks_to_saf(consensus)
counts <- Rsubread::featureCounts(bam_files, annot.ext = saf, isGTFAnnotationFile = FALSE)

}
}
