% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_impacts_functions.R
\name{format_projection}
\alias{format_projection}
\title{format_projection}
\usage{
format_projection(
  data = NULL,
  base_year = NULL,
  gcam_timestep = NULL,
  output_dir = file.path(getwd(), "output")
)
}
\arguments{
\item{data}{Default = NULL. data frame from yield_shock_projection with columns [crop, model, iso, years]}

\item{base_year}{Default = NULL. integer for the base year (for GCAM)}

\item{gcam_timestep}{Default = NULL. integer for the time step of GCAM (Select either 1 or 5 years for GCAM use)}

\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
}
\description{
Format smoothed yield impact projection to wider
}
\keyword{internal}
