% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaea-package.R
\docType{package}
\name{gaea-package}
\alias{gaea}
\alias{gaea-package}
\title{gaea: gaea: An R package to estimate crop yield responses to temperature and precipitation}
\description{
An empirical model to estimate annual crop yield responses to temperature and precipitation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JGCRI/gaea}
  \item Report bugs at \url{https://github.com/JGCRI/gaea/issues}
}

}
\author{
\strong{Maintainer}: Mengqi Zhao \email{mengqi.zhao@pnnl.gov} (\href{https://orcid.org/0000-0001-5385-2758}{ORCID})

Authors:
\itemize{
  \item Stephanie Waldhoff \email{Stephanie.Waldhoff@pnnl.gov}
  \item Claudia Tebaldi \email{claudia.tebaldi@pnnl.gov}
  \item Abigail Snyder \email{abigail.snyder@pnnl.gov}
}

}
\keyword{internal}
