% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaia_cone.R
\name{gaia_cone}
\alias{gaia_cone}
\title{Gaia cone search}
\usage{
gaia_cone(
  ra,
  dec,
  r_arcmin,
  table_name = "gaiadr2.gaia_source",
  output_format = "json",
  row_limit = 50,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\description{
Gaia cone search
}
