% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaia_cone.R
\name{gaia_cone_query}
\alias{gaia_cone_query}
\title{Gaia cone query}
\usage{
gaia_cone_query(
  ra,
  dec,
  r_arcmin,
  table_name,
  output_format,
  row_limit,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\description{
Gaia cone query
}
