% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaia_cone.R
\name{gaia_cone_search}
\alias{gaia_cone_search}
\title{Gaia cone search}
\usage{
gaia_cone_search(
  ra,
  dec,
  radius,
  row_limit = 50L,
  table_name = "gaiadr3.gaia_source",
  cols = NULL,
  verbose = FALSE
)
}
\arguments{
\item{ra}{RA}

\item{dec}{DEC}

\item{radius}{circle radius in degrees}

\item{row_limit}{number of the most nearest sources (default = 50). Set to *-1* for unlimited}

\item{table_name}{gaia table name (default = "gaiadr3.gaia_source")}

\item{cols}{vector of column names (default = NULL)}

\item{verbose}{show ADQL query? (default = FALSE)}
}
\value{
A data frame
}
\description{
Gaia cone search
}
\examples{
\dontrun{
gaia_cone_search(
ra = 95.1669125,
dec = 26.725528,
radius = 15/3600, # 15 arcseconds
table_name = "gaiadr3.gaia_source",
cols = c("source_id", "ra", "ra_error", "dec", "dec_error"),
verbose = TRUE)
}
}
