% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaia_cone.R
\name{gaia_query_sync}
\alias{gaia_query_sync}
\title{Make gaia query}
\usage{
gaia_query_sync(output_format, adql)
}
\arguments{
\item{output_format}{output table format}

\item{adql}{ADQL string}
}
\value{
A `response()` object.
}
\description{
Make gaia query
}
