% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaia_cone.R
\name{gaia_search_by_id}
\alias{gaia_search_by_id}
\title{Search gaia source by source_id}
\usage{
gaia_search_by_id(
  id,
  table_name = "gaiadr3.gaia_source",
  cols = NULL,
  verbose = FALSE
)
}
\arguments{
\item{id}{vector of source_id (bit64::integer64, character)}

\item{table_name}{gaia table name (default = "gaiadr3.gaia_source")}

\item{cols}{vector of column names (default = NULL)}

\item{verbose}{show ADQL query? (default = FALSE)}
}
\value{
A data frame
}
\description{
Search gaia source by source_id
}
\examples{
\dontrun{
sources <- c("3432266562766594688", "3432266562764508032")
sources_int64 <- bit64::as.integer64(sources)
gaia_search_by_id(sources, verbose = TRUE)
gaia_search_by_id(sources_int64, cols = c("source_id", "ra", "ra_error", "dec", "dec_error"))
}
}
