\name{generate_approx_null_hypothesis}
\alias{generate_approx_null_hypothesis}
\title{This function computes the probability distribution of the null hypothesis by an approximation.}
\description{
The probability distribution is computed by performing an approximation of independent random permutations of the rows. 
The null hypothesis is modeled as an histogram with a bin size of 1.
}

\author{Sandro Morganella et al.

Maintainer: S. Morganella <morganellaalx@gmail.com>
}
