\name{generate_null_hypothesis}
\alias{generate_null_hypothesis}
\title{This function computes the probability distribution of the null hypothesis.}
\description{
The probability distribution is computed by performing a number independent random permutation of the rows. 
The null hypothesis is modeled as an histogram with a bin size of 1.
}

\author{Sandro Morganella et al.

Maintainer: S. Morganella <morganellaalx@gmail.com>
}
