% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcam_agprodchange.R
\name{get_cropland_weight}
\alias{get_cropland_weight}
\title{get_cropland_weight}
\usage{
get_cropland_weight(
  gcam_version = "gcam7",
  climate_scenario = NULL,
  output_dir = file.path(getwd(), "output")
)
}
\arguments{
\item{gcam_version}{Default = 'gcam7'. String for the GCAM version. Only support gcam6 and gcam7}

\item{climate_scenario}{Default = NULL. String for climate scenario (e.g., 'ssp245')}

\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
}
\value{
A list of cropland weight within a specific spatial boundary
}
\description{
Calculate the weight of cropland area within the intersected region-glu-country to
cropland area within intersected region-glu
}
\keyword{internal}
