% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_utils.R
\name{get_mirca2000_data}
\alias{get_mirca2000_data}
\title{get_mirca2000_data}
\usage{
get_mirca2000_data(
  download_url =
    "https://zenodo.org/records/7422506/files/harvested_area_grids.zip?download=1",
  file_extension = "zip",
  data_dir = file.path(getwd(), "example")
)
}
\arguments{
\item{download_url}{Default = 'https://zenodo.org/records/7422506/files/harvested_area_grids.zip?download=1'. Link to the downloadable dataset}

\item{file_extension}{Default = 'zip'. String of file extension without "."}

\item{data_dir}{Default = file.path(getwd(), 'example'). Path of desired location to download data}
}
\value{
A string of path that the data is downloaded to
}
\description{
get_mirca2000_data
}
