% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_utils.R
\name{get_nc_time}
\alias{get_nc_time}
\title{get_nc_time}
\usage{
get_nc_time(nc_file = NULL)
}
\arguments{
\item{nc_file}{Default = NULL. String for path to the nc file}
}
\value{
A vector of time series based on the input NetCDF file.
}
\description{
get the netcdf file time series
}
\keyword{internal}
