% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hephaestus_ConvCCA.R
\name{HEPHAESTUS_convCCA_algo}
\alias{HEPHAESTUS_convCCA_algo}
\title{Core ConvCCA Algorithm for Two Datasets}
\usage{
HEPHAESTUS_convCCA_algo(
  X_1,
  X_2,
  tauW_1,
  tauW_2,
  initW_1 = NULL,
  initW_2 = NULL,
  nIter = NULL,
  penalty = "LASSO"
)
}
\description{
Core ConvCCA Algorithm for Two Datasets
}
\keyword{internal}
