% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hephaestus_ConvCCA.R
\name{HEPHAESTUS_multi_convCCA_algo}
\alias{HEPHAESTUS_multi_convCCA_algo}
\title{Core Multi-Dataset ConvCCA Algorithm}
\usage{
HEPHAESTUS_multi_convCCA_algo(
  X,
  tau,
  nIter = 1000,
  penalty = "LASSO",
  initW = NULL
)
}
\description{
Core Multi-Dataset ConvCCA Algorithm
}
\keyword{internal}
