% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hephaestus_RelPMDCCA.R
\name{HEPHAESTUS_multi_relPMDCCA_algo}
\alias{HEPHAESTUS_multi_relPMDCCA_algo}
\title{Core Multi-Dataset RelPMDCCA Algorithm}
\usage{
HEPHAESTUS_multi_relPMDCCA_algo(
  X,
  lambda,
  tau,
  a = 3.7,
  sd.mu = sqrt(2),
  nIter = 1000,
  penalty = "LASSO",
  element_wise = TRUE,
  tau_EN = NULL
)
}
\description{
Core Multi-Dataset RelPMDCCA Algorithm
}
\keyword{internal}
