% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hephaestus_RelPMDCCA.R
\name{HEPHAESTUS_relPMDCCA_algo}
\alias{HEPHAESTUS_relPMDCCA_algo}
\title{Core RelPMDCCA Algorithm for Two Datasets}
\usage{
HEPHAESTUS_relPMDCCA_algo(
  X_1,
  X_2,
  lambda,
  tauW_1,
  tauW_2,
  initW_1 = NULL,
  initW_2 = NULL,
  a = 3.7,
  sd.mu = sqrt(2),
  nIter = 100,
  penalty = "LASSO",
  element_wise = TRUE,
  tau_EN = NULL
)
}
\description{
Core RelPMDCCA Algorithm for Two Datasets
}
\keyword{internal}
