% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hephaestus_ConvCCA.R
\name{HEPHAESTUS_updateWmulti_convCCA}
\alias{HEPHAESTUS_updateWmulti_convCCA}
\title{Update Canonical Vector for Multi-Dataset ConvCCA}
\usage{
HEPHAESTUS_updateWmulti_convCCA(K, w, whichDSet, tau, penalty = "LASSO")
}
\arguments{
\item{K}{List of correlation matrices between datasets}

\item{w}{List of current canonical vectors for all datasets}

\item{whichDSet}{Index of the dataset being updated}

\item{tau}{Sparsity tuning parameter for this dataset}

\item{penalty}{Penalty function: "LASSO" (default) or "SCAD"}
}
\value{
Updated canonical vector for dataset whichDSet
}
\description{
Updates one canonical vector in a multi-dataset setting by
maximizing correlation with all other datasets
}
