% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_utils.R
\name{input_data}
\alias{input_data}
\title{input_data}
\usage{
input_data(
  folder_path = NULL,
  input_file = NULL,
  skip_number = 0,
  quietly = FALSE
)
}
\arguments{
\item{folder_path}{Default = NULL. String for the folder path}

\item{input_file}{Default = NULL. String for the name of the csv file to be read in}

\item{skip_number}{Default = 0. Integer for the number of rows to skip}

\item{quietly}{Default= FALSE. Logical. TRUE to output input data information; FALSE to silent.}
}
\value{
A data table of input CSV file
}
\description{
read in data and output number of rows and columns and variable names
}
\keyword{internal}
