% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_utils.R
\name{iso_replace}
\alias{iso_replace}
\title{iso_replace}
\usage{
iso_replace(d)
}
\arguments{
\item{d}{Data frame with country_name column}
}
\value{
A data frame
}
\description{
ISO replacements for countries with non-standard names (can ad to as they come up)
}
\keyword{internal}
