% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minerva_sCCA_CV.R
\name{MINERVA_cv_convCCA}
\alias{MINERVA_cv_convCCA}
\title{Convenience wrapper for ConvCCA cross-validation}
\usage{
MINERVA_cv_convCCA(...)
}
\arguments{
\item{...}{Arguments passed to MINERVA_cv_scca}
}
\description{
Convenience wrapper for ConvCCA cross-validation
}
