% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minerva_sCCA_CV.R
\name{MINERVA_cv_relPMDCCA}
\alias{MINERVA_cv_relPMDCCA}
\title{Convenience wrapper for RelPMDCCA cross-validation}
\usage{
MINERVA_cv_relPMDCCA(...)
}
\arguments{
\item{...}{Arguments passed to MINERVA_cv_scca}
}
\description{
Convenience wrapper for RelPMDCCA cross-validation
}
