% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minerva_sCCA_CV.R
\name{MINERVA_evaluate_scca}
\alias{MINERVA_evaluate_scca}
\title{Evaluate Sparse CCA Model}
\usage{
MINERVA_evaluate_scca(X_list, W_list, metric = "correlation")
}
\arguments{
\item{X_list}{List of datasets (test data)}

\item{W_list}{List of canonical vectors}

\item{metric}{Metric to compute: "correlation" (default) or "both"
(both computes correlation and sparsity metrics)}
}
\value{
Named list of metrics
}
\description{
Computes evaluation metrics for sCCA results
}
