% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minerva_penalty_functions.R
\name{MINERVA_proximal_scad}
\alias{MINERVA_proximal_scad}
\title{Apply Proximal Operator for SCAD (ADMM Framework)}
\usage{
MINERVA_proximal_scad(omega, mu, c, tau, a = 3.7, old)
}
\arguments{
\item{omega}{Proximal gradient term}

\item{mu}{Step size parameter}

\item{c}{Cross-product term}

\item{tau}{Sparsity tuning parameter}

\item{a}{SCAD shape parameter (default = 3.7)}

\item{old}{Previous iteration value (fallback)}
}
\value{
Updated value
}
\description{
Element-wise proximal operator for SCAD in ADMM.
Used in RelPMDCCA algorithm.
}
