% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcam_agprodchange.R
\name{mirca_to_gcam}
\alias{mirca_to_gcam}
\title{This script includes the following functions:
mirca_to_gcam
get_mirca_cropland
get_cropland_weight
get_weighted_yield_impact
get_agprodchange
gcam_agprodchange
mirca_to_gcam}
\usage{
mirca_to_gcam(gcam_version = "gcam7", climate_scenario = NULL)
}
\arguments{
\item{gcam_version}{Default = 'gcam7'. string for the GCAM version. Only support gcam6 and gcam7}

\item{climate_scenario}{Default = NULL. string for climate scenario (e.g., 'ssp245')}
}
\description{
Mapping between GCAM region, GLU, country
Mapping from mirca crop to GCAM commodities
}
\keyword{internal}
