% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_utils.R
\name{path_check}
\alias{path_check}
\title{path_check}
\usage{
path_check(path = NULL, file_type = NULL)
}
\arguments{
\item{path}{Default = NULL. String for path to the file}

\item{file_type}{Default = NULL. String for extension of the file}
}
\description{
check if file format is correct and file exists
}
\keyword{internal}
