% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_impacts_functions.R
\name{plot_agprodchange}
\alias{plot_agprodchange}
\title{plot_agprodchange}
\usage{
plot_agprodchange(data = NULL, commodity = NULL, output_dir = NULL)
}
\arguments{
\item{data}{Default = NULL. Data frame for the data to plot}

\item{commodity}{Default = NULL. String for GCAM commodity}

\item{output_dir}{Default = NULL. String for path to the output folder}
}
\value{
No return value, called for the side effects of plotting agricultural productivity change by GCAM commodity
}
\description{
Plot projected ag productivity change by GCAM commodity
}
\keyword{internal}
