% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_impacts_functions.R
\name{plot_map}
\alias{plot_map}
\title{plot_map}
\usage{
plot_map(
  data = NULL,
  plot_years = NULL,
  output_dir = file.path(getwd(), "output")
)
}
\arguments{
\item{data}{Default = NULL. Data frame for the data to plot}

\item{plot_years}{Default = NULL. Integer for the years to plot}

\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
}
\value{
No return value, called for the side effects of plotting map of yield shocks at selected years
}
\description{
Plot maps by year, model, scenari, and crop
}
\keyword{internal}
