% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_impacts_functions.R
\name{plot_projection}
\alias{plot_projection}
\title{plot_projection}
\usage{
plot_projection(
  data = NULL,
  climate_model = NULL,
  climate_scenario = NULL,
  crop_name = NULL,
  base_year = NULL,
  output_dir = file.path(getwd(), "output")
)
}
\arguments{
\item{data}{Default = NULL. Data frame for the data to plot}

\item{climate_model}{Default = NULL. String for climate model name}

\item{climate_scenario}{Default = NULL. String for climate scenario (e.g., 'ssp245')}

\item{crop_name}{Default = NULL. String for crop name}

\item{base_year}{Default = NULL. Integer for the base year (for GCAM)}

\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
}
\description{
Figure of (raw) annual predicted values (shows interannual variability)
}
\keyword{internal}
