% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_impacts_functions.R
\name{prep_regression}
\alias{prep_regression}
\title{prep_regression}
\usage{
prep_regression(data = NULL)
}
\arguments{
\item{data}{Default = NULL. Data frame for crop data}
}
\value{
A data frame of formatted historical weather and crop data
}
\description{
Prepare historic dataframe for regression analysis, create all variables that
may be used in model specifications
}
\keyword{internal}
