% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_peak_deg_link.R
\name{print.apollo_deg_peaks}
\alias{print.apollo_deg_peaks}
\title{Print method for apollo_deg_peaks}
\usage{
\method{print}{apollo_deg_peaks}(x, ...)
}
\description{
Print method for apollo_deg_peaks
}
