% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artemis_clustering.R
\name{print.artemis_characterization}
\alias{print.artemis_characterization}
\title{Print method for cluster characterization}
\usage{
\method{print}{artemis_characterization}(x, n = 20, ...)
}
\arguments{
\item{x}{An artemis_characterization object}

\item{n}{Maximum number of items to display}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for cluster characterization
}
