% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artemis_clustering.R
\name{print.artemis_cluster}
\alias{print.artemis_cluster}
\title{Print method for ARTEMIS clustering results}
\usage{
\method{print}{artemis_cluster}(x, ...)
}
\arguments{
\item{x}{An artemis_cluster object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for ARTEMIS clustering results
}
