% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artemis_famd.R
\name{print.artemis_famd}
\alias{print.artemis_famd}
\title{Print method for ARTEMIS FAMD results}
\usage{
\method{print}{artemis_famd}(x, ...)
}
\arguments{
\item{x}{An artemis_famd object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for ARTEMIS FAMD results
}
