% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artemis_discriminative_analysis.R
\name{print.artemis_final_model}
\alias{print.artemis_final_model}
\title{Print method for artemis_final_model}
\usage{
\method{print}{artemis_final_model}(x, ...)
}
\arguments{
\item{x}{An artemis_final_model object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for artemis_final_model
}
