% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artemis_discriminative_analysis.R
\name{print.artemis_group_lasso}
\alias{print.artemis_group_lasso}
\title{Print Method for artemis_group_lasso}
\usage{
\method{print}{artemis_group_lasso}(x, ...)
}
\arguments{
\item{x}{An artemis_group_lasso object}

\item{...}{Additional arguments (unused)}
}
\description{
Print Method for artemis_group_lasso
}
