% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artemis_part_clustering.R
\name{print.artemis_part}
\alias{print.artemis_part}
\title{Print method for artemis_part}
\usage{
\method{print}{artemis_part}(x, ...)
}
\arguments{
\item{x}{An artemis_part object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for artemis_part
}
