% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artemis_timeseries_analysis.R
\name{print.artemis_ts_de}
\alias{print.artemis_ts_de}
\title{Print method for artemis_ts_de}
\usage{
\method{print}{artemis_ts_de}(x, ...)
}
\arguments{
\item{x}{An artemis_ts_de object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for artemis_ts_de
}
