% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artemis_timeseries_analysis.R
\name{print.artemis_ts_norm_limma}
\alias{print.artemis_ts_norm_limma}
\title{Print method for artemis_ts_norm_limma}
\usage{
\method{print}{artemis_ts_norm_limma}(x, ...)
}
\arguments{
\item{x}{An artemis_ts_norm_limma object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for artemis_ts_norm_limma
}
