% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artemis_clustering.R
\name{print.artemis_variable_comparison}
\alias{print.artemis_variable_comparison}
\title{Print method for variable comparison}
\usage{
\method{print}{artemis_variable_comparison}(x, n = 30, ...)
}
\arguments{
\item{x}{An artemis_variable_comparison object}

\item{n}{Maximum number of items to display}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for variable comparison
}
