% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hades_qc_mixed_data.R
\name{print.hades_qc}
\alias{print.hades_qc}
\title{Print method for HADES QC results}
\usage{
\method{print}{hades_qc}(x, show_flagged_only = TRUE, ...)
}
\arguments{
\item{x}{A hades_qc object}

\item{show_flagged_only}{Logical. Only show flagged items. Default = TRUE.}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for HADES QC results
}
