% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hephaestus_MOFA.R
\name{print.hephaestus_mofa}
\alias{print.hephaestus_mofa}
\title{Print method for hephaestus_mofa objects}
\usage{
\method{print}{hephaestus_mofa}(x, ...)
}
\arguments{
\item{x}{A hephaestus_mofa object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for hephaestus_mofa objects
}
