% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artemis_wgcna.R
\name{print.wgcna_data}
\alias{print.wgcna_data}
\title{Print method for wgcna_data objects}
\usage{
\method{print}{wgcna_data}(x, ...)
}
\arguments{
\item{x}{A wgcna_data object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for wgcna_data objects
}
