% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artemis_wgcna.R
\name{print.wgcna_gene_sig}
\alias{print.wgcna_gene_sig}
\title{Print method for wgcna_gene_sig objects}
\usage{
\method{print}{wgcna_gene_sig}(x, ...)
}
\arguments{
\item{x}{A wgcna_gene_sig object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for wgcna_gene_sig objects
}
