% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artemis_wgcna.R
\name{print.wgcna_hubs}
\alias{print.wgcna_hubs}
\title{Print method for wgcna_hubs objects}
\usage{
\method{print}{wgcna_hubs}(x, ...)
}
\arguments{
\item{x}{A wgcna_hubs object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for wgcna_hubs objects
}
