% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artemis_wgcna.R
\name{print.wgcna_trait_cor}
\alias{print.wgcna_trait_cor}
\title{Print method for wgcna_trait_cor objects}
\usage{
\method{print}{wgcna_trait_cor}(x, ...)
}
\arguments{
\item{x}{A wgcna_trait_cor object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for wgcna_trait_cor objects
}
