% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_impacts_functions.R
\name{regression_fixed_effects}
\alias{regression_fixed_effects}
\title{regression_fixed_effects}
\usage{
regression_fixed_effects(
  data = NULL,
  crop_name = NULL,
  formula = NULL,
  output_dir = file.path(getwd(), "output")
)
}
\arguments{
\item{data}{Default = NULL. Data frame for crop data}

\item{crop_name}{Default = NULL. String for crop name}

\item{formula}{Default = NULL. String for regression fomular. Default is waldhoff_formula}

\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
}
\value{
A data frame of historical weather variables used in the regression model
}
\description{
Estimate regression
Regression and estimation of y_hat to compare to historic data.
}
\details{
The regression is based on country fixed effects that control for heterogeneous
unobserved time-invariant influences that vary across countries (e.g., the
distribution of soil quality), and a time-trend that controls for
temporally-varying influences common to all countries (e.g., improved cultivars)
}
\keyword{internal}
