\name{search_peaks_in_regions}
\alias{search_peaks_in_regions}
\title{This function extracts the primary peak within a well-specified region.}
\description{
This function is used as support by the function \code{peel_off}. This function searches and returns the minimum q-value peak (the primary peak) within the region defined by the points \code{start_region} and \code{end_region}.
}
\author{Sandro Morganella et al.

Maintainer: S. Morganella <morganellaalx@gmail.com>
}
