% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_impacts_functions.R
\name{smooth_impacts}
\alias{smooth_impacts}
\title{smooth_impacts}
\usage{
smooth_impacts(
  data = NULL,
  climate_model = NULL,
  climate_scenario = NULL,
  crop_name = NULL,
  base_year = NULL,
  smooth_window = 20
)
}
\arguments{
\item{climate_model}{Default = NULL. string for climate model name}

\item{climate_scenario}{Default = NULL. string for RCP}

\item{crop_name}{Default = NULL. string for crop name}

\item{base_year}{Default = NULL. integer for the base year (for GCAM)}

\item{smooth_window}{Default = 20. integer for smoothing window in years}
}
\description{
Smooth annual impacts using moving average and output certain time step
E.g., 2020_avg = mean(2011:2030)
}
\keyword{internal}
