% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeseq_class.R
\docType{class}
\name{treeseq-class}
\alias{treeseq-class}
\alias{treeseq}
\title{An S4 class to wrap a tskit tree sequence}
\description{
An S4 class to wrap a tskit tree sequence
}
\section{Slots}{

\describe{
\item{\code{treeseq}}{The tree sequence object.}

\item{\code{tree}}{A single marginal tree in the tree sequence.}
}}

