% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeseq_sankoff.R
\name{treeseq_discrete_mpr}
\alias{treeseq_discrete_mpr}
\title{Compute minimum migration costs for genetic ancestors in a tree sequence}
\usage{
treeseq_discrete_mpr(ts, sample_locations, cost_matrix, use_brlen = FALSE)
}
\arguments{
\item{ts}{A \code{treeseq} object.}

\item{sample_locations}{An integer matrix giving the geographic locations
of sampled genomes. The 'node_id' column specifies the node identifier of
of each sample in the tree sequence, and the 'state_id' columns specifies the
geographic state where each sample is found. Note that node identifiers use
a 0-based index while state identifiers use a 1-based index.}

\item{cost_matrix}{A numeric matrix giving the state-to-state migration
costs that will be used by the optimization. All costs should be non-negative
and any non-zero costs on the diagonal will be ignored.}

\item{use_brlen}{A logical specifying whether or not to scale the migration
costs by inverse branch lengths. The default (FALSE) ignores branch lengths
in the optimization.}
}
\description{
Compute minimum migration costs for genetic ancestors in a tree sequence
}
\details{
Returns a numeric vector \code{f} for each ancestral node in the
tree sequence such that \code{f[i]} gives the average (taken over all base
pairs in the ancestor) minimum migration cost needed to explain the
geographic distribution of \emph{all} sample nodes when the ancestor is
located in geographic state \code{i}.
}
