% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeseq_sankoff.R
\name{treeseq_discrete_mpr_ancestry}
\alias{treeseq_discrete_mpr_ancestry}
\title{Spatiotemporal ancestry coefficients}
\usage{
treeseq_discrete_mpr_ancestry(
  ts,
  obj,
  cost_matrix,
  adjacency_matrix,
  times,
  state_sets,
  sample_sets
)
}
\arguments{
\item{ts}{An object of class \code{treeseq}.}

\item{obj}{The result of \code{treeseq_discrete_mpr}.}

\item{cost_matrix}{A numeric matrix giving the state-to-state migration
costs. All costs should be non-negative and any non-zero costs on the 
diagonal will be ignored.}

\item{adjacency_matrix}{A binary matrix specifying geographic connections.
Geographic states \code{i} and \code{j} are connected if the corresponding
entry in \code{adjacency_matrix} is 1.}

\item{times}{A set of times ordered from present to past at which to compute
ancestry coefficients.}

\item{state_sets}{An integer vector with one entry for each geographic
state specifying the set to which that state belongs. Sets must be numbered
beginning from 1.}

\item{sample_sets}{An integer vector with one entry for each sample
specifying the set to which that sample belongs. Sets must be numbered
beginning from 1. A value of 0 may be used to indicate that the sample
belongs to none of the sets and hence is to be excluded from the calculation.}
}
\description{
Spatiotemporal ancestry coefficients
}
