% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeseq_sankoff.R
\name{treeseq_discrete_mpr_edge_history}
\alias{treeseq_discrete_mpr_edge_history}
\title{Sample a migration history for each edge in a tree sequence}
\usage{
treeseq_discrete_mpr_edge_history(
  ts,
  obj,
  cost_matrix,
  adjacency_matrix,
  index1 = TRUE
)
}
\arguments{
\item{ts}{An object of class \code{treeseq}.}

\item{obj}{The result of \code{treeseq_discrete_mpr}.}

\item{cost_matrix}{A numeric matrix giving the state-to-state migration
costs. All costs should be non-negative and any non-zero costs on the 
diagonal will be ignored.}

\item{adjacency_matrix}{A binary matrix specifying geographic connections.
Geographic states \code{i} and \code{j} are connected if the corresponding
entry in \code{adjacency_matrix} is 1.}

\item{index1}{A logical indicating whether or not the returned state
assignments should be indexed from 1 (TRUE) or 0 (FALSE).}
}
\description{
Sample a migration history for each edge in a tree sequence
}
\details{
Samples a minimum cost migration history for each edge in a tree
sequence and returns the history as a data.frame. Each history is a sequence 
of visited geographic states ordered from most recent (present) to least 
recent (past). The first element in the history is the state at the end
of the edge and the last element in the history is the state at the
beginning of the edge. Intermediate elements are the states visited along
the way. If there are no intermediate elements, the edge remained in the
same state for its entire duration.
}
