% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeseq_edges.R
\name{treeseq_drop_edges}
\alias{treeseq_drop_edges}
\title{Drop edges from the tree sequence}
\usage{
treeseq_drop_edges(ts, parent, child)
}
\arguments{
\item{ts}{A \code{treeseq} object.}

\item{parent}{An integer vector of node id's. Edges that emanate from nodes
whose id is in \code{parent} will be dropped.}

\item{child}{An integer vector of node id's. Edges that enter nodes whose 
id is in \code{child} will be dropped.}
}
\value{
A new \code{treeseq} object with the specified edges removed.
}
\description{
Drop edges from the tree sequence
}
