% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeseq_edges.R
\name{treeseq_edges}
\alias{treeseq_edges}
\title{Return the tree sequence edge table}
\usage{
treeseq_edges(ts)
}
\arguments{
\item{ts}{A \code{treeseq} object.}
}
\value{
The tree sequence edge table as a \code{data.frame} object.
}
\description{
Return the tree sequence edge table
}
