% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeseq_individuals.R
\name{treeseq_individuals}
\alias{treeseq_individuals}
\title{Return the tree sequence individuals table}
\usage{
treeseq_individuals(ts)
}
\arguments{
\item{ts}{A \code{treeseq} object.}
}
\value{
The tree sequence individuals table as a \code{matrix} object.
}
\description{
Return the tree sequence individuals table
}
