% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeseq_intervals.R
\name{treeseq_intervals}
\alias{treeseq_intervals}
\title{Return the genomic spans of local trees in a tree sequence}
\usage{
treeseq_intervals(ts)
}
\arguments{
\item{ts}{A \code{treeseq} object.}
}
\value{
A matrix containing a row for each local tree in the tree sequence
specifying the genomic interval over which it relates the samples.
}
\description{
Return the genomic spans of local trees in a tree sequence
}
