% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeseq_load.R
\name{treeseq_load}
\alias{treeseq_load}
\title{Load a tree sequence from a file}
\usage{
treeseq_load(filename)
}
\arguments{
\item{filename}{The name of the file with the tree sequence.}
}
\value{
An object of class \code{treeseq}.
}
\description{
Load a tree sequence from a file
}
