% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeseq_nodes.R
\name{treeseq_nodes}
\alias{treeseq_nodes}
\title{Return the tree sequence node table}
\usage{
treeseq_nodes(ts)
}
\arguments{
\item{ts}{A \code{treeseq} object.}
}
\value{
The tree sequence node table as a \code{data.frame} object.
}
\description{
Return the tree sequence node table
}
