% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeseq_simplify.R
\name{treeseq_simplify}
\alias{treeseq_simplify}
\title{Simplify a tree sequence}
\usage{
treeseq_simplify(
  ts,
  samples,
  node.map = FALSE,
  filter.sites = TRUE,
  filter.populations = TRUE,
  filter.individuals = TRUE,
  no.filter.nodes = FALSE,
  no.update.sample.flags = FALSE,
  reduce.to.site.topology = FALSE,
  keep.unary = FALSE,
  keep.input.roots = FALSE,
  keep.unary.in.individuals = FALSE
)
}
\arguments{
\item{ts}{A \code{treeseq} object.}

\item{samples}{A vector of sample id's.}
}
\value{
A new tree sequence pruned to only include the samples listed in
the \code{samples} parameter.
}
\description{
Simplify a tree sequence
}
