% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_calendars.R
\name{verify_crop}
\alias{verify_crop}
\title{verify_crop}
\usage{
verify_crop(crop_calendar_file = NULL, crop_select = NULL)
}
\arguments{
\item{crop_calendar_file}{Default = NULL. String for the path of the crop calendar file. If crop_cal is provided, crop_select will be set to crops in crop calendar. User provided crop_calendar_file can include any crops MIRCA2000 crops: "wheat", "maize", "rice", "barley", "rye", "millet", "sorghum", "soybean", "sunflower", "root_tuber", "cassava", "sugarcane", "sugarbeet", "oil_palm", "rape_seed", "groundnuts", "pulses", "citrus", "date_palm", "grapes", "cotton", "cocoa", "coffee", "others_perennial", "fodder_grasses", "other_annual"}

\item{crop_select}{Default = NULL. Vector of strings for the selected crops from our database. If NULL, the default crops will be used in the crop calendar: c("cassava", "cotton", "maize", "rice", "root_tuber", "sorghum", "soybean", "sugarbeet", "sugarcane", "sunflower", "wheat"). The additional crops available for selection from our crop calendar database are: "barley", "groundnuts", "millet", "pulses", "rape_seed", "rye"}
}
\value{
A data frame of crop name mapping between MIRCA and SAGE
}
\description{
Identify crops selected by users are available crop types from both SAGE and MIRCA
}
\keyword{internal}
