% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_climate.R
\name{weighted_climate}
\alias{weighted_climate}
\title{weighted_climate}
\usage{
weighted_climate(
  pr_ncdf = NULL,
  tas_ncdf = NULL,
  timestep = "monthly",
  climate_model = "gcm",
  climate_scenario = "rcp",
  time_periods = NULL,
  output_dir = file.path(getwd(), "output"),
  name_append = NULL
)
}
\arguments{
\item{pr_ncdf}{Default = NULL. List of paths for precipitation NetCDF files from ISIMIP}

\item{tas_ncdf}{Default = NULL. List of paths for temperature NetCDF files from ISIMIP}

\item{timestep}{Default = 'monthly'. String for input climate data time step (e.g., 'monthly', 'daily')}

\item{climate_model}{Default = NULL. String for climate model (e.g., 'CanESM5')}

\item{climate_scenario}{Default = NULL. String for climate scenario (e.g., 'ssp245')}

\item{time_periods}{Default = NULL. Vector for years to subset from the climate data. If NULL, use the default climate data period}

\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}

\item{name_append}{Default = NULL. String for name append to the output folder}
}
\value{
No return value, called for the side effects of processing and writing output files
}
\description{
Process standard NetCDF files from ISIMIP to country level climate
based on crop harvested area from MIRCA2000
}
