% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_projections.R
\name{yield_projections}
\alias{yield_projections}
\title{yield_projections}
\usage{
yield_projections(
  use_default_coeff = FALSE,
  climate_model = "gcm",
  climate_scenario = "rcp",
  base_year = NULL,
  start_year = NULL,
  end_year = NULL,
  smooth_window = 20,
  diagnostics = TRUE,
  output_dir = file.path(getwd(), "output")
)
}
\arguments{
\item{use_default_coeff}{Default = FALSE. binary for using default regression coefficients. Set to TRUE will use the default coefficients instead of calculating coefficients from the historical climate data.}

\item{climate_model}{Default = NULL. string for climate model (e.g., 'CanESM5')}

\item{climate_scenario}{Default = NULL. string for climate scenario (e.g., 'ssp245')}

\item{base_year}{Default = NULL. integer for the base year (for GCAM)}

\item{start_year}{Default = NULL. integer for the  start year of the data}

\item{end_year}{Default = NULL. integer for the end year of the data}

\item{smooth_window}{Default = 20. integer for smoothing window in years}

\item{diagnostics}{Default = TRUE. Logical for performing diagnostic plot}

\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
}
\description{
Project yields for future climate scenarios using regression analysis
Using average growing season temperature and precipitation, max and min months
}
