% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_regression.R
\name{yield_regression}
\alias{yield_regression}
\title{yield_regression}
\usage{
yield_regression(
  formula = NULL,
  crop_select = NULL,
  diagnostics = TRUE,
  output_dir = file.path(getwd(), "output")
)
}
\arguments{
\item{formula}{Default = NULL. String for regression formula}

\item{crop_select}{Default = NULL. Vector of strings for the selected crops from our database. If NULL, the default crops will be used in the crop calendar: c("cassava", "cotton", "maize", "rice", "root_tuber", "sorghum", "soybean", "sugarbeet", "sugarcane", "sunflower", "wheat"). The additional crops available for selection from our crop calendar database are: "barley", "groundnuts", "millet", "pulses", "rape_seed", "rye"}

\item{diagnostics}{Default = TRUE. Logical for performing diagnostic plot}

\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
}
\value{
No return value, called for the side effects of processing and writing output files
}
\description{
Yield regressions and create figures
using average growing season temperature and precipitation, max and min months
}
