% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_utils.R
\name{agprodchange_ref}
\alias{agprodchange_ref}
\title{agprodchange_ref}
\usage{
agprodchange_ref(
  gcam_version = "gcam7",
  climate_scenario = NULL,
  gcamdata_dir = NULL
)
}
\arguments{
\item{gcam_version}{Default = 'gcam7'. string for the GCAM version. Only support gcam6 and gcam7}

\item{climate_scenario}{Default = NULL. string for climate scenario (e.g., 'ssp245')}

\item{gcamdata_dir}{Default = NULL. string for directory to the gcamdata folder within the specific GCAM version. The gcamdata need to be run with drake to have the CSV outputs beforehand.}
}
\description{
Get the reference agricultural productivity change based GCAM version
}
\keyword{internal}
