% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_impacts_functions.R
\name{clean_yield}
\alias{clean_yield}
\title{clean_yield}
\usage{
clean_yield(fao_yield = NULL, fao_to_mirca = NULL)
}
\arguments{
\item{fao_yield}{Default = NULL. Data frame for the fao yield table}

\item{fao_to_mirca}{Default = NULL. Data frame for the fao to mirca crop mapping}
}
\value{
A data frame of formatted FAO yield data
}
\description{
Function to clean FAO yield data
}
\keyword{internal}
