% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_impacts_functions.R
\name{climate_impact}
\alias{climate_impact}
\title{climate_impact}
\usage{
climate_impact(
  use_default_coeff = FALSE,
  climate_model = NULL,
  climate_scenario = NULL,
  crop_name = NULL,
  base_year = NULL,
  start_year = NULL,
  end_year = NULL,
  output_dir = file.path(getwd(), "output")
)
}
\arguments{
\item{use_default_coeff}{Default = FALSE. binary for using default regression coefficients. Set to TRUE will use the default coefficients instead of calculating coefficients from the historical climate data.}

\item{climate_model}{Default = NULL. string for climate model name}

\item{climate_scenario}{Default = NULL. string for RCP}

\item{crop_name}{Default = NULL. string for crop name}

\item{base_year}{Default = NULL. integer for the base year (for GCAM)}

\item{start_year}{Default = NULL. integer for the  start year of the data}

\item{end_year}{Default = NULL. integer for the end year of the data}

\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
}
\description{
Estimate climate impacts (ratio future to each base year, then average)
}
\keyword{internal}
