% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{co2_projection}
\alias{co2_projection}
\title{co2_projection}
\format{
R data frame
}
\source{
GCAM7 reference CO2 concentration
}
\usage{
co2_projection
}
\description{
co2_projection
}
\examples{
\dontrun{
 library(gaia);
 co2_projection <- gaia::co2_projection
}
}
\keyword{datasets}
