% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coef_default}
\alias{coef_default}
\title{coef_default}
\format{
R data frame
}
\source{
Regression coefficients derived from the default historical climate data
}
\usage{
coef_default
}
\description{
coef_default
}
\examples{
\dontrun{
 library(gaia);
 coef_default <- gaia::coef_default
}
}
\keyword{datasets}
