% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_utils.R
\name{colname_replace}
\alias{colname_replace}
\title{colname_replace}
\usage{
colname_replace(d, x, y)
}
\arguments{
\item{d}{Data frame}

\item{x}{String to be replaced}

\item{y}{Replacing string}
}
\value{
A data frame
}
\description{
Shorter command to replace column names
}
\keyword{internal}
