% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_id}
\alias{country_id}
\title{country_id}
\format{
R data frame
}
\source{
country ID and name mapping
}
\usage{
country_id
}
\description{
Mapping of country ID, country name, and country ISO codes. This table can be used to identify countries based on their IDs in weighted climate files generated by the `gaia::weighted_climate` function. For mode details, see the [User Guide](https://jgcri.github.io/gaia/articles/vignette.html#weighted_climate).
}
\examples{
\dontrun{
 library(gaia);
 country_id <- gaia::country_id
}
}
\keyword{datasets}
