% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_calendars.R
\name{crop_calendars}
\alias{crop_calendars}
\title{crop_calendars}
\usage{
crop_calendars(
  crop_calendar_file = NULL,
  crop_select = NULL,
  output_dir = file.path(getwd(), "output")
)
}
\arguments{
\item{crop_calendar_file}{Default = NULL. String for the path of the crop calendar file. If crop_calendar_file is provided, crop_select will be set to crops in crop calendar. User provided crop_calendar_file can include any crops MIRCA2000 crops: "wheat", "maize", "rice", "barley", "rye", "millet", "sorghum", "soybean", "sunflower", "root_tuber", "cassava", "sugarcane", "sugarbeet", "oil_palm", "rape_seed", "groundnuts", "pulses", "citrus", "date_palm", "grapes", "cotton", "cocoa", "coffee", "others_perennial", "fodder_grasses", "others_annual"}

\item{crop_select}{Default = NULL. Vector of strings for the selected crops from our database. If NULL, the default crops will be used in the crop calendar: c("cassava", "cotton", "maize", "rice", "root_tuber", "sorghum", "soybean", "sugarbeet", "sugarcane", "sunflower", "wheat"). The additional crops available for selection from our crop calendar database are: "barley", "groundnuts", "millet", "pulses", "rape_seed", "rye"}

\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
}
\value{
A data frame of crop calendars
}
\description{
Generate planting months for each crop and country using SAGE global crop planting and harvesting dates data [Sacks et al., 2010](https://doi.org/10.1111/j.1466-8238.2010.00551.x).
}
