% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_impacts_functions.R
\name{crop_month}
\alias{crop_month}
\title{crop_month}
\usage{
crop_month(climate_data = NULL, crop_name = NULL, crop_calendar = NULL)
}
\arguments{
\item{climate_data}{Default = NULL. Data table for climate data}

\item{crop_name}{Default = NULL. String for crop name}

\item{crop_calendar}{Default = NULL. Data table for crop calendar}
}
\value{
A data frame of crop growing season
}
\description{
Function to estimate growing season for each crop and country (SAGE db)
}
\keyword{internal}
