% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_impacts_functions.R
\name{data_trans}
\alias{data_trans}
\title{data_trans}
\usage{
data_trans(
  data = NULL,
  climate_model = NULL,
  climate_scenario = NULL,
  crop_name = NULL,
  output_dir = NULL,
  co2_proj = co2_projection
)
}
\arguments{
\item{data}{Default = NULL. data table for crop data created by crop_month}

\item{climate_model}{Default = NULL. string for climate model}

\item{climate_scenario}{Default = NULL. string for climate scenario}

\item{crop_name}{Default = NULL. string for crop name}

\item{output_dir}{Default = NULL. string for path to the output folder}

\item{co2_proj}{Default = NULL. data table for future CO2 concentration [year, co2_conc]}
}
\description{
Create dataframe with future monthly weather variables and iso codes for
projecting yields using fitted regression model
}
\keyword{internal}
