% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eleuthia_logging.R
\name{.log_env}
\alias{.log_env}
\title{Eleuthia - Logging Functions}
\usage{
.log_env
}
\description{
Functions for managing console output logging to files.
Duplicates console output to both screen and log file (does not redirect).
}
