% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaia-package.R
\docType{package}
\name{gaia-package}
\alias{gaia}
\alias{gaia-package}
\title{gaia: gaia: An R package to estimate crop yield responses to temperature and precipitation}
\description{
An empirical model to estimate annual crop yield responses to temperature and precipitation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JGCRI/gaia}
  \item Report bugs at \url{https://github.com/JGCRI/gaia/issues}
}

}
\author{
\strong{Maintainer}: Mengqi Zhao \email{mengqi.zhao@pnnl.gov} (\href{https://orcid.org/0000-0001-5385-2758}{ORCID})

Authors:
\itemize{
  \item Stephanie T. Morris \email{Stephanie.morris@pnnl.gov} (\href{https://orcid.org/0000-0002-8073-0868}{ORCID})
  \item Claudia Tebaldi \email{claudia.tebaldi@pnnl.gov} (\href{https://orcid.org/0000-0001-9233-8903}{ORCID})
  \item Abigail Snyder \email{abigail.snyder@pnnl.gov} (\href{https://orcid.org/0000-0002-9034-9948}{ORCID})
}

}
\keyword{internal}
