% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAIA-package.R
\docType{package}
\name{GAIA-package}
\alias{GAIA}
\alias{GAIA-package}
\title{GAIA: Multi-Omics Integration Toolkit for CUT&TAG, ATAC-seq, and RNA-seq}
\description{
Modular pipeline for multi-omics analysis including quality control, preprocessing, annotation, enrichment, integration, machine learning, statistical analysis, visualization, and reporting. Named after the subordinate functions from Horizon Zero Dawn.
}
\author{
\strong{Maintainer}: Yohan Lefol \email{yohan.lefol@gmail.com}

Authors:
\itemize{
  \item Yohan Lefol \email{yohan.lefol@gmail.com}
}

}
\keyword{internal}
