% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcam_agprodchange.R
\name{gcam_agprodchange}
\alias{gcam_agprodchange}
\title{gcam_agprodchange}
\usage{
gcam_agprodchange(
  data = NULL,
  gcamdata_dir = NULL,
  climate_model = "gcm",
  climate_scenario = "rcp",
  member = "member",
  bias_adj = "ba",
  gcam_version = "gcam7",
  gcam_timestep = 5,
  cfe = "no-cfe",
  base_year = 2015,
  diagnostics = TRUE,
  output_dir = file.path(getwd(), "output")
)
}
\arguments{
\item{data}{Default = NULL. Output data frame from function yield_shock_projection, or similar format of data}

\item{gcamdata_dir}{Default = NULL. String for directory to the gcamdata folder within the specific GCAM version. The gcamdata need to be run with drake to have the CSV outputs beforehand.}

\item{climate_model}{Default = 'gcm'. String for climate model name (e.g., 'CanESM5')}

\item{climate_scenario}{Default = 'rcp'. String for climate scenario name (e.g., 'ssp245')}

\item{member}{Default = 'member'. String for the ensemble member name}

\item{bias_adj}{Default = 'ba'. String for the dataset used for climate data bias adjustment}

\item{gcam_version}{Default = 'gcam7'. String for the GCAM version. Only support gcam6 and gcam7}

\item{gcam_timestep}{Default = 5. Integer for the time step of GCAM (Select either 1 or 5 years for GCAM use)}

\item{cfe}{Default = 'no-cfe'. String for whether the yield impact formula implimented CO2 fertilization effect.}

\item{base_year}{Default = 2015. Integer for the base year (for GCAM)}

\item{diagnostics}{Default = TRUE. Logical for performing diagnostic plot}

\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
}
\value{
A data frame of formatted agricultural productivity change for GCAM
}
\description{
Remaps country-level yield shocks to GCAM-required spatial scales (e.g., region, basin, and intersections), based on harvested areas, and aggregates crops to GCAM commodities. This function applies the projected shocks to GCAM scenario agricultural productivity growth rates (the unit used to project future yields in GCAM) and creates ready-to-use XML outputs for GCAM.
}
