% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcam_agprodchange.R
\name{gcam_agprodchange}
\alias{gcam_agprodchange}
\title{gcam_agprodchange}
\usage{
gcam_agprodchange(data = NULL, output_dir = NULL)
}
\arguments{
\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
}
\description{
Map country level yield impacts to GCAM region-GLU level
Calculate agricultural productivity change by region GLU
}
