% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcam_agprodchange.R
\name{get_agprodchange}
\alias{get_agprodchange}
\title{-----------------------------------------------------------------------------
get_agprodchange}
\usage{
get_agprodchange(data = NULL, from_year = NULL, to_year = NULL)
}
\arguments{
\item{data}{Default = NULL. output data from function yield_projections, or similar format of data}

\item{gcam_version}{Default = 'gcam7'. string for the GCAM version. Only support gcam6 and gcam7}

\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
}
\description{
Calculate agricultural productivity change based on yield impact multiplier and no impact yield
}
\keyword{internal}
