% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcam_agprodchange.R
\name{get_agprodchange}
\alias{get_agprodchange}
\title{-----------------------------------------------------------------------------
get_agprodchange}
\usage{
get_agprodchange(data = NULL, from_year = NULL, to_year = NULL)
}
\arguments{
\item{data}{Default = NULL. output data from function yield_projections, or similar format of data}

\item{from_year}{Default = NULL. integer for 'from' year to calculate agricultural productivity change}

\item{to_year}{Default = NULL. integer for 'to' year to calculate agricultural productivity change}
}
\description{
Calculate agricultural productivity change based on yield impact multiplier and no impact yield
}
\keyword{internal}
