% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_utils.R
\name{get_example_data}
\alias{get_example_data}
\title{get_example_data}
\usage{
get_example_data(
  download_url = "",
  file_extension = "zip",
  data_dir = file.path(getwd(), "example"),
  check_exist = TRUE
)
}
\arguments{
\item{download_url}{Link to the downloadable dataset}

\item{file_extension}{Default = 'zip'. String of file extension without "."}

\item{data_dir}{Default= file.path(getwd(), 'example'). Path of desired location to download data}

\item{check_exist}{Default = TRUE. Binary to check if the example data already exists. If TRUE, then function gives error if the downloaded folder already exist}
}
\description{
get_example_data
}
