% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_utils.R
\name{get_example_data}
\alias{get_example_data}
\title{get_example_data}
\usage{
get_example_data(
  download_url = "",
  file_extension = "zip",
  data_dir = file.path(getwd(), "example")
)
}
\arguments{
\item{download_url}{Link to the downloadable dataset}

\item{file_extension}{String of file extension without "."}

\item{data_dir}{Path of desired location to download data}
}
\description{
get_example_data
}
