% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcam_agprodchange.R
\name{get_mirca_cropland}
\alias{get_mirca_cropland}
\title{-----------------------------------------------------------------------------
get_mirca_cropland}
\usage{
get_mirca_cropland(raster_brick = NULL, mapping = NULL)
}
\arguments{
\item{raster_brick}{Default = NULL. raster brick for the mirca cropland}

\item{mapping}{Default = NULL. data table of 0.5 degree 67420 grid cells with lat and lon of global land area}
}
\description{
Calculate cropland fraction within country and glu intersection
}
\keyword{internal}
