% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcam_agprodchange.R
\name{get_weighted_yield_impact}
\alias{get_weighted_yield_impact}
\title{-----------------------------------------------------------------------------
get_weighted_yield_impact}
\usage{
get_weighted_yield_impact(
  data = NULL,
  gcam_version = "gcam7",
  diagnostics = TRUE,
  output_dir = file.path(getwd(), "output")
)
}
\arguments{
\item{data}{Default = NULL. output data from function yield_projections, or similar format of data}

\item{gcam_version}{Default = 'gcam7'. string for the GCAM version. Only support gcam6 and gcam7}

\item{diagnostics}{Default = TRUE. Logical for performing diagnostic plot}

\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
}
\description{
Calculate weighted yield impact by region-glu scale for each GCAM commodity
}
\keyword{internal}
