% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_utils.R
\name{merge_data}
\alias{merge_data}
\title{merge_data}
\usage{
merge_data(d1, d2, x1, x2)
}
\arguments{
\item{d1}{Data frame}

\item{d2}{Data frame}

\item{x1}{String for column name in d1}

\item{x2}{String for column name in d2}
}
\value{
A data frame
}
\description{
Merge columns from two data frames using two variables
}
\keyword{internal}
