% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_impacts_functions.R
\name{plot_fit}
\alias{plot_fit}
\title{plot_fit}
\usage{
plot_fit(
  data = NULL,
  crop_name = NULL,
  output_dir = file.path(getwd(), "output")
)
}
\arguments{
\item{data}{Default = NULL. data frame for crop data}

\item{crop_name}{Default = NULL. string for crop name}

\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
}
\description{
Figure of fitted regression estimates compared to FAO yields
}
\keyword{internal}
