% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_impacts_functions.R
\name{plot_map}
\alias{plot_map}
\title{plot_map}
\usage{
plot_map(data = NULL, plot_years = NULL, output_dir = NULL)
}
\arguments{
\item{data}{Default = NULL. data frame for the data to plot}

\item{plot_years}{Default = NULL. integer for the years to plot}

\item{output_dir}{Default = NULL. string for path to the output folder}
}
\description{
Plot maps by year, model, scenari, and crop
}
\keyword{internal}
