% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_impacts_functions.R
\name{plot_yield_impact}
\alias{plot_yield_impact}
\title{plot_yield_impact}
\usage{
plot_yield_impact(
  data = NULL,
  commodity = NULL,
  crop_type = NULL,
  output_dir = NULL
)
}
\arguments{
\item{data}{Default = NULL. Data frame for the data to plot}

\item{commodity}{Default = NULL. String for GCAM commodity}

\item{crop_type}{Default = NULL. String for mirca crop type}

\item{output_dir}{Default = NULL. String for path to the output folder}
}
\value{
No return value, called for the side effects of plotting projected yield impact by GCAM commodity
}
\description{
Plot projected yield impact by GCAM commodity
}
\keyword{internal}
