% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_impacts_functions.R
\name{smooth_impacts}
\alias{smooth_impacts}
\title{smooth_impacts}
\usage{
smooth_impacts(
  data = NULL,
  climate_model = NULL,
  climate_scenario = NULL,
  crop_name = NULL,
  base_year = NULL,
  start_year = NULL,
  end_year = NULL,
  smooth_window = 20,
  output_dir = file.path(getwd(), "output")
)
}
\arguments{
\item{climate_model}{Default = NULL. string for climate model name}

\item{climate_scenario}{Default = NULL. string for climate scenario (e.g., 'ssp245')}

\item{crop_name}{Default = NULL. string for crop name}

\item{base_year}{Default = NULL. integer for the base year (for GCAM)}

\item{start_year}{Default = NULL. integer for the  start year of the data}

\item{end_year}{Default = NULL. integer for the end year of the data}

\item{smooth_window}{Default = 20. integer for smoothing window in years}

\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
}
\description{
Smooth annual impacts using moving average and output certain time step
E.g., 2020_avg = mean(2011:2030)
}
\keyword{internal}
