% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeseq_edges.R
\name{treeseq_num_edges}
\alias{treeseq_num_edges}
\title{Return the number of edges in a tree sequence}
\usage{
treeseq_num_edges(ts)
}
\arguments{
\item{ts}{A \code{treeseq} object.}
}
\description{
Return the number of edges in a tree sequence
}
