% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeseq_nodes.R
\name{treeseq_num_nodes}
\alias{treeseq_num_nodes}
\title{Return the number of nodes in a tree sequence}
\usage{
treeseq_num_nodes(ts)
}
\arguments{
\item{ts}{A \code{treeseq} object.}
}
\description{
Return the number of nodes in a tree sequence
}
