% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeseq_populations.R
\name{treeseq_populations}
\alias{treeseq_populations}
\title{Return the tree sequence populations table}
\usage{
treeseq_populations(ts)
}
\arguments{
\item{ts}{A \code{treeseq} object.}
}
\value{
The tree sequence populations table as a \code{matrix} object.
}
\description{
Return the tree sequence populations table
}
