% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeseq_write.R
\name{treeseq_write}
\alias{treeseq_write}
\title{Write a tree sequence to a file}
\usage{
treeseq_write(ts, filename)
}
\arguments{
\item{ts}{A \code{treeseq} object.}

\item{filename}{The name of the file to write to.}
}
\description{
Write a tree sequence to a file
}
