% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_impacts_functions.R
\name{weather_agg}
\alias{weather_agg}
\title{weather_agg}
\usage{
weather_agg(file_precip = NULL, file_temp = NULL, crop_name = NULL)
}
\arguments{
\item{file_precip}{Default = NULL. string for path to the precipitation file}

\item{file_temp}{Default = NULL. string for path to the temperature file}

\item{crop_name}{Default = NULL. string for crop name}
}
\description{
Function to aggregate the cleaned temperature and precipitation data for each crop
}
\keyword{internal}
