% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_climate.R
\name{weighted_climate}
\alias{weighted_climate}
\title{weighted_climate}
\usage{
weighted_climate(
  pr_files = NULL,
  tas_files = NULL,
  timestep = "monthly",
  climate_model = "gcm",
  climate_scenario = "rcp",
  time_periods = NULL,
  output_dir = file.path(getwd(), "output")
)
}
\arguments{
\item{pr_files}{Default = NULL. list of paths for precipitation}

\item{tas_files}{Default = NULL. list of paths for temperature}

\item{timestep}{Default = 'monthly'. string for input climate data time step (e.g., 'monthly', 'daily')}

\item{climate_model}{Default = NULL. string for climate model (e.g., 'CanESM5')}

\item{climate_scenario}{Default = NULL. string for climate scenario (e.g., 'ssp245')}

\item{time_periods}{Default = NULL. vector for years to input in the output}

\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
}
\description{
Process standard NetCDF files from ISIMIP to country level climate
based on crop harvested area from MIRCA2000
}
