% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_regression.R
\name{yield_regression}
\alias{yield_regression}
\title{yield_regression}
\usage{
yield_regression(
  formula = NULL,
  diagnostics = TRUE,
  output_dir = file.path(getwd(), "output")
)
}
\arguments{
\item{formula}{Default = NULL. String for regression formula}

\item{diagnostics}{Default = TRUE. Logical for performing diagnostic plot}

\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
}
\value{
No return value, called for the side effects of processing and writing output files
}
\description{
Yield regressions and create figures
using average growing season temperature and precipitation, max and min months
}
