% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_regression.R
\name{yield_regression}
\alias{yield_regression}
\title{yield_regression}
\usage{
yield_regression(
  formula = NULL,
  diagnostics = TRUE,
  output_dir = file.path(getwd(), "output")
)
}
\arguments{
\item{formula}{Default = NULL. String for regression formula}

\item{diagnostics}{Default = TRUE. Logical for performing diagnostic plot}

\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
}
\value{
No return value, called for the side effects of processing and writing output files
}
\description{
Performs regression analysis fitted with historical annual crop yields, growing season monthly temperature and precipitation, CO2 concentrations, and GDP per capita. The default econometric model applied in gaia is from \href{https://www.doi.org/10.1088/1748-9326/abadcb}{Waldhoff et al., (2020)}. User can specify alternative formulas that are consistent with the data processed in `data_aggregation`.
}
