% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_shock_projection.R
\name{yield_shock_projection}
\alias{yield_shock_projection}
\title{yield_shock_projection}
\usage{
yield_shock_projection(
  use_default_coeff = FALSE,
  climate_model = "gcm",
  climate_scenario = "rcp",
  base_year = 2015,
  start_year = NULL,
  end_year = NULL,
  gcam_timestep = 5,
  smooth_window = 20,
  diagnostics = TRUE,
  output_dir = file.path(getwd(), "output")
)
}
\arguments{
\item{use_default_coeff}{Default = FALSE. Binary for using default regression coefficients. Set to TRUE will use the default coefficients instead of calculating coefficients from the historical climate data.}

\item{climate_model}{Default = NULL. String for climate model (e.g., 'CanESM5')}

\item{climate_scenario}{Default = NULL. String for climate scenario (e.g., 'ssp245')}

\item{base_year}{Default = 2015. Integer for the base year (for GCAM)}

\item{start_year}{Default = NULL. Integer for the  start year of the data}

\item{end_year}{Default = NULL. Integer for the end year of the data}

\item{gcam_timestep}{Default = 5. Integer for the time step of GCAM (Select either 1 or 5 years for GCAM use)}

\item{smooth_window}{Default = 20. Integer for smoothing window in years}

\item{diagnostics}{Default = TRUE. Logical for performing diagnostic plot}

\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
}
\value{
A data frame of formatted smoothed annual crop yield shocks under climate impacts
}
\description{
Project yields for future climate scenarios using regression analysis
Using average growing season temperature and precipitation, max and min months
}
