% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_impacts_functions.R
\name{z_estimate}
\alias{z_estimate}
\title{z_estimate}
\usage{
z_estimate(
  use_default_coeff = FALSE,
  climate_model = NULL,
  climate_scenario = NULL,
  crop_name = NULL,
  output_dir = file.path(getwd(), "output")
)
}
\arguments{
\item{use_default_coeff}{Default = FALSE. Binary for using default regression coefficients. Set to TRUE will use the default coefficients instead of calculating coefficients from the historical climate data.}

\item{climate_model}{Default = NULL. String for climate model name}

\item{climate_scenario}{Default = NULL. String for climate scenario (e.g., 'ssp245')}

\item{crop_name}{Default = NULL. String for crop name}

\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
}
\value{
A data frame of variables in the regression model
}
\description{
Estimate Z_t = ln(Y_t) w/o constants
}
\keyword{internal}
