% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_impacts_functions.R
\name{z_estimate}
\alias{z_estimate}
\title{z_estimate}
\usage{
z_estimate(
  climate_model = NULL,
  climate_scenario = NULL,
  crop_name = NULL,
  output_dir = file.path(getwd(), "output")
)
}
\arguments{
\item{climate_model}{Default = NULL. string for climate model name}

\item{climate_scenario}{Default = NULL. string for RCP}

\item{crop_name}{Default = NULL. string for crop name}

\item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
}
\description{
Estimate Z_t = ln(Y_t) w/o constants
}
\keyword{internal}
